[include-fixer] Work around partial names in both directions.
We already handled the case where we had a nested name specifier where
parts from the beginning don't get a callback, also handle the case
where the end doesn't get a callback. This happens with function calls
in unknown namespaces.

The way we do this for now based on character data is a bit of a hack, we
may need to refine this later or fix clang to produce different callbacks.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@269029 91177308-0d34-0410-b5e6-96231b3b80d8