[include-fixer] Also look up prefixes of queries.
This is used to find nested classes. For a nested name foo::bar::qux we
will first look up foo::bar::qux, then foo::bar, then foo unless we find
a result. This is used to support nested classes which are not part of
the index but can only be used if the header for the parent context is
included.

Differential Revision: http://reviews.llvm.org/D20372

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@269956 91177308-0d34-0410-b5e6-96231b3b80d8