[clang-rename] Use checktime when reloading vim buffer after applying clang-rename
After applying `clang-rename` to a vim buffer (using `clang-rename.py` as part
of the vim integration) the buffer gets reloaded using `bufdo`. This solution is
suboptimal, since syntax highlighting is turned off for performance reasons and
never turned on, after all changes to the source file have been applied.

A better solution to this is using `checktime`. It is exactly designed for this
kind of task and doesn't have the syntax highlighting issue.

Patch by Kai Wolf!

Reviewers: omtcyfz

Differential Revision: https://reviews.llvm.org/D24791


git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@282388 91177308-0d34-0410-b5e6-96231b3b80d8