[clang-tidy] Partly rewrite readability-simplify-boolean-expr using RAV
The check was using AST matchers in a very inefficient manner. By rewriting the
BinaryOperator-related parts using RAV, the check was sped up by a factor of
up to 10000 on some files (mostly, generated code using binary operators in
tables), but also significantly sped up for regular large files.
As a side effect, the code became clearer and more readable.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@303081 91177308-0d34-0410-b5e6-96231b3b80d8