[clangd] Don't try pthread, just use thread_local. Reverts r323949.
The pthread solution here breaks standalone builds, which don't have the
relevant cmake magic for feature-detection.

The original reason for trying pthread was fear of libgcc without
support for thread_local (e.g. on the clang-x86_64-linux-selfhost-modules bot).
However the earliest supported GCC is 4.8, and this has __cxa_thread_atexit.

This will probably break that bot, it's not running a supported GCC and needs
to be upgraded. I'll try to find out how to do this.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@324351 91177308-0d34-0410-b5e6-96231b3b80d8