[clang-tidy] Remove google-runtime-member-string-references
This is triggering on a pattern that's both too broad (const
std::string& members can be used safely) and too narrow (std::string is
not the only class with this problem). It has a very low true positive
rate, just remove it until we find a better solution for dangling string
references.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@329292 91177308-0d34-0410-b5e6-96231b3b80d8