[clangd] Drop fixes if replying with tweaks resulted in an error
This should not happen in normal operation, as it implies that the diagnostics
with some available fixes were produced but the AST is invalid.
Moreover, the code had an error: always returned code actions ignoring the
SupportsCodeAction capability and writing a test for this is impossible,
since this can only happen due to programmer's error rather than invalid inputs.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@352624 91177308-0d34-0410-b5e6-96231b3b80d8