Merging r366687: ------------------------------------------------------------------------ r366687 | aaronballman | 2019-07-22 15:22:08 +0200 (Mon, 22 Jul 2019) | 3 lines
Update documentation for all CERT checks that correspond to a recommendation.

CERT removed their C++ secure coding recommendations from public view and so the links within that documentation are stale. This updates various pieces of documentation to make this more clear, and to help add substance where our docs deferred to CERT's wiki.
------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/branches/release_90@366710 91177308-0d34-0410-b5e6-96231b3b80d8