[clang-doc] Add index in each info html file
An index structure is created while generating the output file for each
info. This structure is parsed to JSON and written to a file in the
output directory. The html for the index is not rendered by clang-doc. A
Javascript file is included in the output directory, this will the JSON
file and insert HTML elements into the file.

Differential Revision: https://reviews.llvm.org/D65690

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@368070 91177308-0d34-0410-b5e6-96231b3b80d8